fix milliseconds to timeval conversion for SO_RCVTIMEO
On my platform (Mac OS X 10.8.5; Darwin 12.5.0) and possibly others,
SO_RCVTIMEO will reject timevals with a tv_usec value of 1000000 or
greater with an EDOM errno ("Numerical argument out of domain").
This causes the "seed" executable to fail, as it sets a timeout of
1000 milliseconds.

This patch correctly calculates the timeval, and allows the seed to
run.